Move mapnik renderer instantiation to mapnik factory
