+++ lib/windshaft/renderers/mapnik/index.js
@@ -1,1 +1,2 @@
 module.exports.adaptor = require('./adaptor');
+module.exports.factory = require('./factory');
