+++ lib/windshaft/renderers/render_cache.js
@@ -6,8 +6,6 @@ var   _                     = require('underscore')
     , MapnikRenderer = require('./mapnik')
     , TorqueRenderer = require('./torque');
 
-require('tilelive-mapnik').registerProtocols(tilelive);
-
 module.exports = function(options, mml_store, map_store, mapnik_opts) {
 
     // Configure bases for cache keys suitable for string interpolation
@@ -218,123 +216,26 @@ module.exports = function(options, mml_store, map_store, mapnik_opts) {
         return dbParams;
     };
 
-    me.mapConfigToMMLBuilderConfig = function(mapConfig) {
-        var cfg = mapConfig.obj();
-        var sql = [];
-        var style = [];
-        var geom_columns = [];
-        var extra_ds_opts = [];
-        var interactivity = [];
-        var style_version = [];
-        for ( var i=0; i<cfg.layers.length; ++i ) {
-          var lyr = cfg.layers[i];
-          if ( mapConfig.layerType(i) != 'mapnik' ) continue;
-          if ( ! lyr.hasOwnProperty('options') )
-            throw new Error("Missing options from layer " + i + " of layergroup config");
-          var lyropt = lyr.options;
-          if ( ! lyropt.hasOwnProperty('sql') )
-            throw new Error("Missing sql for layer " + i + " options");
-          // Wrap SQL requests in mapnik format if sent
-          sql.push( "(" + lyropt.sql.replace(/;\s*$/, '') + ") as cdbq");
-          if ( ! lyropt.hasOwnProperty('cartocss') )
-            throw new Error("Missing cartocss for layer " + i + " options");
-          style.push(lyropt.cartocss);
-          if ( ! lyropt.hasOwnProperty('cartocss_version') ) {
-            throw new Error("Missing cartocss_version for layer " + i + " options");
-          }
-          style_version.push(lyropt.cartocss_version);
-          // NOTE: interactivity used to be a string as of version 1.0.0
-          if ( _.isArray(lyropt.interactivity) ) {
-            lyropt.interactivity = lyropt.interactivity.join(',');
-          } 
-          interactivity.push(lyropt.interactivity);
-          if (lyropt['geom_column']) {
-              geom_columns[i] = {
-                  type: lyropt['geom_type'], // possibly undefined, grainstore allows it
-                  name: lyropt['geom_column']
-              };
-          }
-          extra_opt = {};
-          if ( lyropt.hasOwnProperty('raster_band') ) {
-            extra_opt['band'] = lyropt['raster_band'];
-          }
-          extra_ds_opts.push( extra_opt );
-        }
-        if ( ! sql.length ) throw new Error("No 'mapnik' layers in MapConfig");
-        var opts = {
-            sql: sql,
-            style: style,
-            style_version: style_version,
-            interactivity: interactivity,
-            ttl: 0,
-            extra_ds_opts: extra_ds_opts
-        };
-        if (geom_columns.length) {
-            opts.gcols = geom_columns;
-        }
-
-        return opts;
-    };
+    var mapnikFactory = new MapnikRenderer.factory(mml_store, {
+        mapnik_opts: mapnik_opts,
+        createKeyFn: me.createKey
+    });
 
     // controls the instantiation of mapnik renderer objects from mapnik XML
     me.makeRendererMapnik = function(req, callback){
-        var that = this;
         var params = _.extend({}, req.params);
-        var mml_builder; 
 
         // returns a tilelive renderer by:
         // 1. generating or retrieving mapnik XML
         // 2. configuring a full mml document with mapnik XML, interactivity and other elements set
         Step(
             function getConfig(){
-              if ( ! params.token ) return null;
-              map_store.load(params.token, this);
+                if ( ! params.token ) return null;
+                map_store.load(params.token, this);
             },
-            function initBuilder(err, mapConfig) {
-              if (err) throw err;
-              if ( params.token ) {
-                delete params.interactivity; // will be rewritten from MapConfig
-                params = _.defaults(params, that.mapConfigToMMLBuilderConfig(mapConfig));
-                delete params.token;
-              }
-              // create an mapnik mml builder object
-              mml_builder = mml_store.mml_builder(params, this);
-            },
-            function generateXML(err){
+            function initRenderer(err, mapConfig) {
                 if (err) throw err;
-                mml_builder.toXML(this);
-            },
-            function loadMapnik(err, data){
-                if (!err && req.profiler) {
-                    req.profiler.done('generateXML');
-                }
-
-                if (err) throw err;
-
-                var query = {
-                    // TODO: document what `base` is
-                    base: that.createKey(params) + '/xia', 
-                    metatile: that.mapnik_opts.metatile,
-                    bufferSize: that.mapnik_opts.bufferSize,
-                    autoLoadFonts: false,
-                    internal_cache: false
-                };
-
-
-                // build full document to pass to tilelive
-                var uri = {
-                    query: query,
-                    protocol: 'mapnik:',
-                    slashes: true,
-                    xml: data,
-                    strict: !!params.strict, // force boolean
-                    mml: {
-                        format: params.format // this seems to be useless
-                    }
-                };
-
-                // hand off to tilelive to create a renderer
-                tilelive.load(uri, this);
+                mapnikFactory.getRenderer(mapConfig, params, params.format, params.layer, this);
             },
             function makeAdaptor(err, source){
                 if (!err && req.profiler) {
